# api/docs/API/ExpenseLit_API_OpenAPI_Swagger.yaml
openapi: 3.0.3
info:
  title: ExpenseLit API
  description: |
    API RESTful para gerenciamento completo de finanças pessoais com sistema robusto de permissões e criptografia de dados sensíveis.
    
    ## Funcionalidades Principais
    - ✅ Gerenciamento de contas bancárias e cartões de crédito
    - ✅ Controle de despesas e receitas por categorias
    - ✅ Sistema de empréstimos e transferências
    - ✅ Cadastro unificado de membros (usuários, credores, beneficiários)
    - ✅ Criptografia automática de dados sensíveis (CVV, números de conta/cartão)
    - ✅ Sistema robusto de permissões baseado no Django
    - ✅ Autenticação JWT com refresh tokens
    
    ## Segurança
    - **Criptografia Fernet**: CVVs e números são automaticamente criptografados
    - **JWT Authentication**: Tokens seguros com refresh
    - **Sistema de Permissões**: Controle granular baseado no Django
    - **Validações Robustas**: Entrada de dados sanitizada
    - **Middleware de Segurança**: Headers de segurança automáticos
    
  version: 1.0.0
  contact:
    name: ExpenseLit API Support
    email: tarcisio.ribeiro.1840@hotmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8002/api/v1
    description: Servidor de Desenvolvimento
  - url: https://api.expenselit.com/v1
    description: Servidor de Produção
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de autenticação
  schemas:
...
    BaseModel:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        uuid:
          type: string
          format: uuid
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        is_deleted:
          type: boolean
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true

    # Account Schema
...
    Member:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          required:
            - name
            - document
            - phone
            - sex
          properties:
            name:
              type: string
              maxLength: 200
            document:
              type: string
              maxLength: 200
              description: CPF/CNPJ único
            phone:
              type: string
              maxLength: 200
            email:
              type: string
              maxLength: 200
              nullable: true
            sex:
              type: string
              enum: [M, F]
            user:
              type: integer
              nullable: true
              description: ID do usuário do sistema (OneToOneField)
            is_user:
              type: boolean
              readOnly: true
              description: True se vinculado a um usuário (calculado)
            is_creditor:
              type: boolean
              default: true
            is_benefited:
              type: boolean
              default: true
            active:
              type: boolean
              default: true
            birth_date:
              type: string
              format: date
              nullable: true
            address:
              type: string
              nullable: true
            profile_photo:
              type: string
              format: binary
              nullable: true
            emergency_contact:
              type: string
              maxLength: 200
              nullable: true
            monthly_income:
              type: string
              format: decimal
              nullable: true
            occupation:
              type: string
              maxLength: 200
              nullable: true
            notes:
              type: string
              nullable: true
            age:
              type: integer
              readOnly: true
              description: Idade calculada baseada na data de nascimento

    # Loan Schema
...
paths:
...
  /members/{id}/:
    get:
      tags:
        - Members
      summary: Obter Membro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes do membro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'

    put:
      tags:
        - Members
      summary: Atualizar Membro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '200':
          description: Membro atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'

    delete:
      tags:
        - Members
      summary: Excluir Membro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Membro excluído com sucesso

  # Loans endpoints
...